requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.7
ecdsa==0.19.2
fastapi==0.135.2
h11==0.16.0
httptools==0.7.1
idna==3.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
